WIP - Refactored local/parachain into modules
